pyproject.toml
[project]
name = "wordling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
]
